[taier-refactor]migrate useCalcTableScroll
