+++ taier-ui/src/components/customHooks/index.ts
@@ -17,3 +17,4 @@
  */
 
 export { default as useEnv } from './useEnv';
+export { default as useCalcTableScroll } from './useCalcTableScroll'
